+++ src/PSRule.Rules.Azure/rules/Azure.Resource.Rule.ps1
@@ -21,3 +21,16 @@ Rule 'Azure.Resource.UseTags' -If { SupportsTags } -Tag @{ release = 'GA' } {
 Rule 'Azure.Resource.AllowedRegions' -If { ($Null -ne $Configuration.Azure_AllowedRegions) -and ($Configuration.Azure_AllowedRegions.Length -gt 0) -and (SupportsRegions) } -Tag @{ release = 'GA' } {
     IsAllowedRegion
 }
+
+# Synopsis: Use Resource Group naming requirements
+Rule 'Azure.ResourceGroup.Name' -Type 'Microsoft.Resources/resourceGroups' -Tag @{ release = 'GA' } {
+    # https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules#microsoftresources
+
+    # Between 1 and 90 characters long
+    $Assert.GreaterOrEqual($TargetObject, 'Name', 1)
+    $Assert.LessOrEqual($TargetObject, 'Name', 90)
+
+    # Alphanumerics, underscores, parentheses, hyphens, periods
+    # Can't end with period.
+    Match 'Name' '^[-\w\._\(\)]*[-\w_\(\)]$'
+}