+++ src/PSRule.Rules.Azure/rules/Azure.VirtualNetwork.Rule.ps1
@@ -69,7 +69,74 @@ Rule 'Azure.VNET.PeerState' -If { (HasPeerNetwork) } -Tag @{ release = 'GA' } {
     }
 }
 
+# Synopsis: Use VNET naming requirements
+Rule 'Azure.VNET.Name' -Type 'Microsoft.Network/virtualNetworks' -Tag @{ release = 'GA' } {
+    # https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules#microsoftnetwork
+
+    # Between 2 and 64 characters long
+    $Assert.GreaterOrEqual($TargetObject, 'Name', 2)
+    $Assert.LessOrEqual($TargetObject, 'Name', 64)
+
+    # Alphanumerics, underscores, periods, and hyphens
+    # Start with alphanumeric
+    # End alphanumeric or underscore
+    Match 'Name' '^[A-Za-z0-9](-|\w|\.){0,}\w$'
+}
+
+# Synopsis: Use standard subnets names
+Rule 'Azure.VNET.SubnetName' -Type 'Microsoft.Network/virtualNetworks', 'Microsoft.Network/virtualNetworks/subnets' -Tag @{ release = 'GA' } {
+    # https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules#microsoftnetwork
+
+    if ($PSRule.TargetType -eq 'Microsoft.Network/virtualNetworks') {
+        foreach ($subnet in $TargetObject.Properties.subnets) {
+            if ($subnet.name -eq 'GatewaySubnet') {
+                $Assert.Pass();
+            }
+            else {
+                # Between 1 and 80 characters long
+                $Assert.GreaterOrEqual($subnet, 'Name', 1)
+                $Assert.LessOrEqual($subnet, 'Name', 80)
+
+                # Alphanumerics, underscores, periods, and hyphens.
+                # Start with alphanumeric. End alphanumeric or underscore.
+                $subnet | Match 'Name' '^[A-Za-z0-9]((-|\.)*\w){0,79}$'
+            }
+        }
+    }
+    elseif ($PSRule.TargetType -eq 'Microsoft.Network/virtualNetworks/subnets') {
+        if ($PSRule.TargetName -eq 'GatewaySubnet') {
+            $Assert.Pass();
+        }
+        else {
+            # Between 1 and 80 characters long
+            $Assert.GreaterOrEqual($TargetObject, 'Name', 1)
+            $Assert.LessOrEqual($TargetObject, 'Name', 80)
+
+            # Alphanumerics, underscores, periods, and hyphens.
+            # Start with alphanumeric. End alphanumeric or underscore.
+            $TargetObject | Match 'Name' '^[A-Za-z0-9]((-|\.)*\w){0,79}$'
+        }
+    }
+}
+
 #endregion Virtual Network
+
+#region Route tables
+
+# Synopsis: Use standard route table names
+Rule 'Azure.Route.Name' -Type 'Microsoft.Network/routeTables' -Tag @{ release = 'GA' } {
+    # https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules#microsoftnetwork
+
+    # Between 1 and 80 characters long
+    $Assert.GreaterOrEqual($TargetObject, 'Name', 1)
+    $Assert.LessOrEqual($TargetObject, 'Name', 80)
+
+    # Alphanumerics, underscores, periods, and hyphens.
+    # Start with alphanumeric. End alphanumeric or underscore.
+    Match 'Name' '^[A-Za-z0-9]((-|\.)*\w){0,79}$'
+}
+
+#endregion Route tables
 
 #region Network Security Group
 
@@ -116,6 +183,19 @@ Rule 'Azure.NSG.Associated' -Type 'Microsoft.Network/networkSecurityGroups' -If 
     Reason $LocalizedData.ResourceNotAssociated
     $Assert.HasFieldValue($TargetObject, 'Properties.subnets').Result -or
         $Assert.HasFieldValue($TargetObject, 'Properties.networkInterfaces').Result
+}
+
+# Synopsis: Use standard network security group names
+Rule 'Azure.NSG.Name' -Type 'Microsoft.Network/networkSecurityGroups' -Tag @{ release = 'GA' } {
+    # https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules#microsoftnetwork
+
+    # Between 1 and 80 characters long
+    $Assert.GreaterOrEqual($TargetObject, 'Name', 1)
+    $Assert.LessOrEqual($TargetObject, 'Name', 80)
+
+    # Alphanumerics, underscores, periods, and hyphens.
+    # Start with alphanumeric. End alphanumeric or underscore.
+    Match 'Name' '^[A-Za-z0-9]((-|\.)*\w){0,79}$'
 }
 
 #endregion Network Security Group
@@ -176,6 +256,23 @@ Rule 'Azure.AppGw.WAFRules' -If { (IsAppGwWAF) } -Tag @{ release = 'GA' } {
 
 #endregion Application Gateway
 
+#region Public IP
+
+# Synopsis: Use standard public IP address names
+Rule 'Azure.PublicIP.Name' -Type 'Microsoft.Network/publicIPAddresses' -Tag @{ release = 'GA' } {
+    # https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules#microsoftnetwork
+
+    # Between 1 and 80 characters long
+    $Assert.GreaterOrEqual($TargetObject, 'Name', 1)
+    $Assert.LessOrEqual($TargetObject, 'Name', 80)
+
+    # Alphanumerics, underscores, periods, and hyphens.
+    # Start with alphanumeric. End alphanumeric or underscore.
+    Match 'Name' '^[\w][-\w_\.]*[\w_]$'
+}
+
+#endregion Public IP
+
 #region Load Balancer
 
 # Synopsis: Use specific network probe
@@ -196,6 +293,19 @@ Rule 'Azure.LB.Probe' -Type 'Microsoft.Network/loadBalancers' -Tag @{ release = 
     }
 }
 
+# Synopsis: Use standard load balancer names
+Rule 'Azure.LB.Name' -Type 'Microsoft.Network/loadBalancers' -Tag @{ release = 'GA' } {
+    # https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules#microsoftnetwork
+
+    # Between 1 and 80 characters long
+    $Assert.GreaterOrEqual($TargetObject, 'Name', 1)
+    $Assert.LessOrEqual($TargetObject, 'Name', 80)
+
+    # Alphanumerics, underscores, periods, and hyphens.
+    # Start with alphanumeric. End alphanumeric or underscore.
+    Match 'Name' '^[A-Za-z0-9]((-|\.)*\w){0,79}$'
+}
+
 #endregion Load Balancer
 
 #region Azure Firewall
@@ -207,84 +317,50 @@ Rule 'Azure.Firewall.Mode' -Type 'Microsoft.Network/azureFirewalls' -Tag @{ rele
 
 #endregion Azure Firewall
 
-#region Front Door
-
-# Synopsis: Front Door instance should be enabled
-Rule 'Azure.FrontDoor.State' -Type 'Microsoft.Network/frontDoors' -Tag @{ release = 'GA' } {
-    $Assert.HasFieldValue($TargetObject, 'Properties.enabledState', 'Enabled');
-}
-
-# Synopsis: Use a minimum of TLS 1.2
-Rule 'Azure.FrontDoor.MinTLS' -Type 'Microsoft.Network/frontDoors', 'Microsoft.Network/frontDoors/frontendEndpoints' -Tag @{ release = 'GA' } {
-    $endpoints = @($TargetObject);
-    if ($PSRule.TargetType -eq 'Microsoft.Network/frontDoors') {
-        $endpoints = @($TargetObject.Properties.frontendEndpoints);
-    }
-    foreach ($endpoint in $endpoints) {
-        $Assert.HasDefaultValue($endpoint, 'properties.customHttpsConfiguration.minimumTlsVersion', '1.2');
-    }
-    # properties.frontendEndpoints[].properties.customHttpsConfiguration.minimumTlsVersion
-}
-
-# Synopsis: Use diagnostics to audit Front Door access
-Rule 'Azure.FrontDoor.Logs' -Type 'Microsoft.Network/frontDoors' -Tag @{ release = 'GA' } {
-    Reason $LocalizedData.DiagnosticSettingsNotConfigured;
-    $diagnostics = @(GetSubResources -ResourceType 'microsoft.insights/diagnosticSettings', 'Microsoft.Network/frontDoors/providers/diagnosticSettings');
-    $logCategories = @($diagnostics | ForEach-Object {
-        foreach ($log in $_.Properties.logs) {
-            if ($log.category -eq 'FrontdoorAccessLog' -and $log.enabled -eq $True) {
-                $log;
-            }
-        }
-    });
-    $Null -ne $logCategories -and $logCategories.Length -gt 0;
-}
-
-# Synopsis: Enable WAF policy of each endpoint
-Rule 'Azure.FrontDoor.UseWAF' -Type 'Microsoft.Network/frontDoors', 'Microsoft.Network/frontDoors/frontendEndpoints' -Tag @{ release = 'GA' } {
-    $endpoints = @($TargetObject);
-    if ($PSRule.TargetType -eq 'Microsoft.Network/frontDoors') {
-        $endpoints = @($TargetObject.Properties.frontendEndpoints);
-    }
-    foreach ($endpoint in $endpoints) {
-        $Assert.HasFieldValue($endpoint, 'properties.webApplicationFirewallPolicyLink.id');
-    }
-}
-
-# Synopsis: Use Front Door WAF policy in prevention mode
-Rule 'Azure.FrontDoor.WAF.Mode' -Type 'Microsoft.Network/frontdoorwebapplicationfirewallpolicies' -Tag @{ release = 'GA' } {
-    $Assert.HasFieldValue($TargetObject, 'Properties.policySettings.mode', 'Prevention');
-}
-
-# Synopsis: Enable Front Door WAF policy
-Rule 'Azure.FrontDoor.WAF.Enabled' -Type 'Microsoft.Network/frontdoorwebapplicationfirewallpolicies' -Tag @{ release = 'GA' } {
-    $Assert.HasFieldValue($TargetObject, 'Properties.policySettings.enabledState', 'Enabled');
-}
-
-#endregion Front Door
-
-#region VPN Gateways
+#region Virtual Network Gateway
 
 # Synopsis: Migrate from legacy VPN gateway SKUs
-Rule 'Azure.VPNGateway.LegacySKU' -Type 'Microsoft.Network/virtualNetworkGateways' -If { IsVPNGateway } -Tag @{ release = 'GA' } {
+Rule 'Azure.VNG.VPNLegacySKU' -Type 'Microsoft.Network/virtualNetworkGateways' -If { IsVPNGateway } -Tag @{ release = 'GA' } {
     Within 'Properties.sku.name' -Not 'Basic', 'Standard', 'HighPerformance';
 }
 
 # Synopsis: Use Active-Active configuration
-Rule 'Azure.VPNGateway.ActiveActive' -Type 'Microsoft.Network/virtualNetworkGateways' -If { IsVPNGateway } -Tag @{ release = 'GA' } {
+Rule 'Azure.VNG.VPNActiveActive' -Type 'Microsoft.Network/virtualNetworkGateways' -If { IsVPNGateway } -Tag @{ release = 'GA' } {
     $Assert.HasFieldValue($TargetObject, 'Properties.activeActive', $True);
 }
 
-#endregion VPN Gateways
-
-#region ExpressRoute
-
 # Synopsis: Migrate from legacy ExpressRoute gateway SKUs
-Rule 'Azure.ERGateway.LegacySKU' -Type 'Microsoft.Network/virtualNetworkGateways' -If { IsERGateway } -Tag @{ release = 'GA' } {
+Rule 'Azure.VNG.ERLegacySKU' -Type 'Microsoft.Network/virtualNetworkGateways' -If { IsERGateway } -Tag @{ release = 'GA' } {
     Within 'Properties.sku.name' -Not 'Basic';
 }
 
-#endregion ExpressRoute
+# Synopsis: Use standard virtual network gateway names
+Rule 'Azure.VNG.Name' -Type 'Microsoft.Network/virtualNetworkGateways' -Tag @{ release = 'GA' } {
+    # https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules#microsoftnetwork
+
+    # Between 1 and 80 characters long
+    $Assert.GreaterOrEqual($TargetObject, 'Name', 1)
+    $Assert.LessOrEqual($TargetObject, 'Name', 80)
+
+    # Alphanumerics, underscores, periods, and hyphens.
+    # Start with alphanumeric. End alphanumeric or underscore.
+    Match 'Name' '^[A-Za-z0-9]((-|\.)*\w){0,79}$'
+}
+
+# Synopsis: Use standard virtual networks gateway connection names
+Rule 'Azure.VNG.ConnectionName' -Type 'Microsoft.Network/connections' -Tag @{ release = 'GA' } {
+    # https://docs.microsoft.com/en-us/azure/azure-resource-manager/management/resource-name-rules#microsoftnetwork
+
+    # Between 1 and 80 characters long
+    $Assert.GreaterOrEqual($TargetObject, 'Name', 1)
+    $Assert.LessOrEqual($TargetObject, 'Name', 80)
+
+    # Alphanumerics, underscores, periods, and hyphens.
+    # Start with alphanumeric. End alphanumeric or underscore.
+    Match 'Name' '^[A-Za-z0-9]((-|\.)*\w){0,79}$'
+}
+
+#endregion Virtual Network Gateway
 
 #region Helper functions