+++ CHANGELOG.md
@@ -3,9 +3,34 @@
 ## Unreleased
 
 - New rules:
-  - ExpressRoute Gateway:
+  - Azure Kubernetes Service:
+    - Check AKS cluster name requirements. [#373](https://github.com/Microsoft/PSRule.Rules.Azure/issues/373)
+    - Check AKS cluster DNS prefix requirements. [#373](https://github.com/Microsoft/PSRule.Rules.Azure/issues/373)
+  - Container Registry:
+    - Check registry name requirements. [#373](https://github.com/Microsoft/PSRule.Rules.Azure/issues/373)
+  - Front Door:
+    - Check Front Door name requirements. [#373](https://github.com/Microsoft/PSRule.Rules.Azure/issues/373)
+  - Load Balancer:
+    - Check Load Balancer name requirements. [#373](https://github.com/Microsoft/PSRule.Rules.Azure/issues/373)
+  - Network Security Group:
+    - Check NSG name requirements. [#373](https://github.com/Microsoft/PSRule.Rules.Azure/issues/373)
+  - Public IP:
+    - Check Public IP name requirements. [#373](https://github.com/Microsoft/PSRule.Rules.Azure/issues/373)
+  - Resource Group:
+    - Check Resource Group name requirements. [#373](https://github.com/Microsoft/PSRule.Rules.Azure/issues/373)
+  - Route table
+    - Check Route table name requirements. [#373](https://github.com/Microsoft/PSRule.Rules.Azure/issues/373)
+  - SignalR Service:
+    - Check SignalR Service name requirements. [#373](https://github.com/Microsoft/PSRule.Rules.Azure/issues/373)
+  - Storage:
+    - Check Storage Account name requirements. [#373](https://github.com/Microsoft/PSRule.Rules.Azure/issues/373)
+  - Virtual Network:
+    - Check VNET name requirements. [#373](https://github.com/Microsoft/PSRule.Rules.Azure/issues/373)
+    - Check VNET subnet name requirements. [#373](https://github.com/Microsoft/PSRule.Rules.Azure/issues/373)
+  - Virtual Network Gateway:
+    - Check VNG name requirements. [#373](https://github.com/Microsoft/PSRule.Rules.Azure/issues/373)
+    - Check VNG connection name requirements. [#373](https://github.com/Microsoft/PSRule.Rules.Azure/issues/373)
     - Check ExpressRoute Gateway uses current SKU. [#369](https://github.com/Microsoft/PSRule.Rules.Azure/issues/369)
-  - VPN Gateway:
     - Check VPN Gateway uses current SKU. [#370](https://github.com/Microsoft/PSRule.Rules.Azure/issues/370)
     - Check VPN Gateway uses active-active configuration. [#371](https://github.com/Microsoft/PSRule.Rules.Azure/issues/371)